add numeric sensor for mysql

+++ dqoai/src/integration-test/java/ai/dqo/mysql/column/numeric/MysqlColumnNumericValueBelowMinValueCountSensorParametersSpecIntegrationTest.java
@@ -66,7 +66,7 @@ public class MysqlColumnNumericValueBelowMinValueCountSensorParametersSpecIntegr
         Table resultTable = sensorResult.getResultTable();
         Assertions.assertEquals(1, resultTable.rowCount());
         Assertions.assertEquals("actual_value", resultTable.column(0).name());
-        Assertions.assertEquals(15L, resultTable.column(0).get(0));
+        Assertions.assertEquals(15, (double) resultTable.column(0).get(0));
     }
 
     @Test
@@ -81,7 +81,7 @@ public class MysqlColumnNumericValueBelowMinValueCountSensorParametersSpecIntegr
         Table resultTable = sensorResult.getResultTable();
         Assertions.assertEquals(1, resultTable.rowCount());
         Assertions.assertEquals("actual_value", resultTable.column(0).name());
-        Assertions.assertEquals(15L, resultTable.column(0).get(0));
+        Assertions.assertEquals(15, (double) resultTable.column(0).get(0));
     }
 
     @Test
@@ -96,7 +96,7 @@ public class MysqlColumnNumericValueBelowMinValueCountSensorParametersSpecIntegr
         Table resultTable = sensorResult.getResultTable();
         Assertions.assertEquals(1, resultTable.rowCount());
         Assertions.assertEquals("actual_value", resultTable.column(0).name());
-        Assertions.assertEquals(15L, resultTable.column(0).get(0));
+        Assertions.assertEquals(15, (double) resultTable.column(0).get(0));
     }
 
     @Test
@@ -111,7 +111,7 @@ public class MysqlColumnNumericValueBelowMinValueCountSensorParametersSpecIntegr
         Table resultTable = sensorResult.getResultTable();
         Assertions.assertEquals(6, resultTable.rowCount());
         Assertions.assertEquals("actual_value", resultTable.column(0).name());
-        Assertions.assertEquals(0L, resultTable.column(0).get(0));
+        Assertions.assertEquals(0, (double) resultTable.column(0).get(0));
     }
 
     @Test
@@ -126,6 +126,6 @@ public class MysqlColumnNumericValueBelowMinValueCountSensorParametersSpecIntegr
         Table resultTable = sensorResult.getResultTable();
         Assertions.assertEquals(6, resultTable.rowCount());
         Assertions.assertEquals("actual_value", resultTable.column(0).name());
-        Assertions.assertEquals(0L, resultTable.column(0).get(0));
+        Assertions.assertEquals(0, (double) resultTable.column(0).get(0));
     }
 }
